# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ada21Tech AI-Powered Student Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admins/login:
    post:
      operationId: loginAdmin
      tags: [admins]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions with optional filters
      parameters:
        - in: query
          name: subjectId
          schema:
            type: integer
        - in: query
          name: grade
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a new question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /feedback/{sessionId}:
    get:
      operationId: getFeedback
      tags: [feedback]
      summary: Get AI feedback for a session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiFeedback"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feedback/{sessionId}/generate:
    post:
      operationId: generateFeedback
      tags: [feedback]
      summary: Generate or regenerate AI feedback for a session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiFeedback"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
...
    AdminAuthResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [admin]
      required:
        - token
        - role
...
    Question:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        grade:
          type: string
        text:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        explanation:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subjectId
        - subjectName
        - grade
        - text
        - optionA
        - optionB
        - optionC
        - optionD
        - correctOption
        - explanation
        - difficulty
        - createdAt
    CreateQuestionBody:
      type: object
      properties:
        subjectId:
          type: integer
        grade:
          type: string
        text:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        explanation:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
      required:
        - subjectId
        - grade
        - text
        - optionA
        - optionB
        - optionC
        - optionD
        - correctOption
        - explanation
    UpdateQuestionBody:
      type: object
      properties:
        text:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        explanation:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
...
    AiFeedback:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        summary:
          type: ["string", "null"]
        strengths:
          type: ["string", "null"]
        weaknesses:
          type: ["string", "null"]
        recommendations:
          type: ["string", "null"]
        questionFeedback:
          type: ["string", "null"]
        generatedAt:
          type: ["string", "null"]
      required:
        - id
        - sessionId
        - summary
        - strengths
        - weaknesses
        - recommendations
        - questionFeedback
        - generatedAt
...